Update deployment workflow to include build and deploy steps (#40)
* chore: Update deployment workflow to include build and deploy steps

* chore: Update deployment workflow to include build and deploy steps

* chore: Remove unnecessary required flag in workflow inputs

* chore: Refactor deployment workflows to remove unnecessary install step

* chore: Update node version to 20 and cache yarn in build workflow

* chore: Add success step to pull_request workflow

This commit adds a new step called "success" to the pull_request workflow. This step runs on the ubuntu-latest runner and is triggered after the "build" step. The purpose of this step is to echo a success message and output a variable called "success" with the value "true". This step is designed to trigger auto merge if all dependencies are met.