chore: Update npm dependencies and devDependencies
This commit updates the npm dependencies and devDependencies in the package.json file. It adds the "@testing-library/jest-dom" and "@testing-library/react" packages as devDependencies, and updates the "eslint" and "eslint-config-next" packages to their latest versions. It also adds the "jest" and "jest-environment-jsdom" packages as devDependencies. These updates and additions ensure that the project is using the latest stable versions of the required packages and improves the development and testing experience.